+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+* Added newest and oldest
+* Added replaceItem and replaceFirstItemWhere methods
+
 ## 0.0.2
 
 * Added scrollToBottom and jumpToBottom